requirements.txt
accelerate==1.3.0
aim==3.27.0
aim-ui==3.27.0
aimrecords==0.0.7
aimrocks==0.5.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
base58==2.0.1
beautifulsoup4==4.13.3
bitarray==3.0.0
blinker==1.9.0
boto3==1.36.15
botocore==1.36.15
cachetools==5.5.1
cbor==1.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
colbert-ai==0.2.21
cryptography==44.0.0
dataclasses-json==0.6.7
datasets==2.19.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.0
filelock==3.17.0
FlagEmbedding==1.3.3
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.3.1
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
huggingface-hub==0.28.1
idna==3.10
ijson==3.3.0
inscriptis==2.5.3
ir_datasets==0.5.9
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.43
langchain-milvus==0.1.4
langchain-openai==0.1.20
langchain-text-splitters==0.2.4
langserve==0.2.3
langsmith==0.1.147
lit==18.1.8
lxml==5.3.0
lz4==4.4.3
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
milvus-lite==2.4.11
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
neo4j==5.24.0
networkx==3.4.2
ninja==1.11.1.3
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==1.61.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
peft==0.14.0
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyarrow==19.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymilvus==2.5.4
pyproject-toml==0.0.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
RestrictedPython==8.0
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.11.2
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sse-starlette==2.1.2
starlette==0.38.6
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.19.1
toml==0.10.2
torch==2.0.1
tqdm==4.67.1
transformers==4.44.2
trec-car-tools==2.6
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
unlzw3==0.2.3
urllib3==2.3.0
uvicorn==0.23.2
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
websockets==14.2
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
zlib-state==0.1.9
